requirements.txt
langgraph>=0.0.15
langchain-google-genai>=0.0.8
langchain-core>=0.1.20
streamlit>=1.31.0
torch>=2.2.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
plotly>=5.18.0
python-dotenv>=1.0.1
